requirements.txt
aiohttp
discord.py
dnspython
jishaku
motor
python-dotenv
